requirements.txt
Flask
python-dotenv
google-genai
numpy
gunicorn